% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost.R
\name{create_S_from_gbtregressor}
\alias{create_S_from_gbtregressor}
\title{Iterative Smoother Matrix Construction from XGBoost Regressor}
\usage{
create_S_from_gbtregressor(
  model,
  leaf_indices_train,
  leaf_indices_test,
  base_score,
  save_output,
  output_dir
)
}
\arguments{
\item{model}{An object of class \code{xgb.Booster}.}

\item{leaf_indices_train}{Matrix. Leaf indices for the training data (rows=samples, cols=trees).}

\item{leaf_indices_test}{Matrix. Leaf indices for the test data.}

\item{base_score}{Numeric or character. The model’s initial prediction (0 or "mean").}

\item{save_output}{Logical. If TRUE, saves intermediate matrices to disk.}

\item{output_dir}{Character. Directory path to save intermediate RDS files.}
}
\value{
A list containing:
\itemize{
\item \code{S_train}: The smoother matrix for the training set (n_train x n_train).
\item \code{S_test}: The smoother matrix for the test set (n_test x n_train).
}
}
\description{
Iterates through all trees in an XGBoost model to
construct the final smoother matrix.
}
\keyword{internal}
