% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml_with_smoother.R
\name{dml_inference}
\alias{dml_inference}
\title{Results and influence functions for linear double ML estimators.}
\usage{
dml_inference(psi.a, psi.b)
}
\arguments{
\item{psi.a}{psi.a component of linear Neyman orthogonal score}

\item{psi.b}{psi.a component of linear Neyman orthogonal score}
}
\value{
List of three components:
\itemize{
\item \code{TaPa} Matrix storing estimate, standard error, t-value and p-value of target parameter of each repetition in the rows
\item \code{IF} Matrix storing the influence function of the target parameter of each repetition in the columns
\item \code{score} Array of dimension n_reps x N x 3 where the matrix in the first dimension stores the score in the first,
psi.a in the second, and psi.b in the third column
}
}
\description{
Results and influence functions for linear double ML estimators.
}
\keyword{internal}
