% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml_with_smoother.R
\name{dml_with_smoother}
\alias{dml_with_smoother}
\title{This function runs Double ML estimators with outcome smoothers.}
\usage{
dml_with_smoother(
  Y,
  D,
  X,
  Z = NULL,
  estimators = c("PLR", "PLR_IV", "AIPW_ATE", "Wald_AIPW"),
  smoother = "honest_forest",
  n_cf_folds = 5,
  n_reps = 1,
  ...
)
}
\arguments{
\item{Y}{Numeric vector containing the outcome variable.}

\item{D}{Optional binary treatment variable.}

\item{X}{Covariate matrix with N rows and p columns.}

\item{Z}{Optional binary instrumental variable.}

\item{estimators}{String (vector) indicating which estimators should be run.
Current menu: c("PLR","PLR_IV","AIPW_ATE","Wald_AIPW")}

\item{smoother}{Indicate which smoother to be used for nuisance parameter estimation.
Currently only available option \code{"honest_forest"} from the \pkg{grf} package.}

\item{n_cf_folds}{Number of cross-fitting folds. Default is 5.}

\item{n_reps}{Number of repetitions of cross-fitting. Default is 1.}

\item{...}{Options to be passed to smoothers.}
}
\value{
A list with three entries:
\itemize{
\item \code{results}: a list storing the results, influence functions, and score functions of each estimator
\item \code{NuPa.hat}: a list storing the estimated nuisance parameters and the outcome smoother matrices
}
}
\description{
Existing Double ML implementations are too general to easily extract smoother matrices
required to be compatible with the get_forest_weights() method. This motivates yet
another Double ML implementation.
}
