% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoubleML.R
\name{get_DoubleML_smoother}
\alias{get_DoubleML_smoother}
\title{Create smoother matrix for \code{DoubleML} models}
\usage{
get_DoubleML_smoother(
  fold_models,
  test_ids_list,
  train_ids_list,
  data,
  x_cols,
  y_col,
  subset = NULL
)
}
\arguments{
\item{fold_models}{List of fold-specific models from the \code{DoubleML} object.}

\item{test_ids_list}{List of test indices for each fold.}

\item{train_ids_list}{List of training indices for each fold.}

\item{data}{A numeric matrix with the variables used to construct
the \code{DoubleMLData} object.}

\item{x_cols}{Character vector of covariate column names.}

\item{y_col}{Character string specifying the outcome column name.}

\item{subset}{Logical vector indicating which observations to use for
extracting smoother weights. If not provided, all observations are used.}
}
\value{
An \eqn{N \times N} smoother matrix.
}
\description{
Computes an \eqn{N \times N} smoother matrix for the corresponding nuisance parameter
by looping over the fold-specific models trained within \code{DoubleML}.
}
\references{
Knaus, M. C. (2024). Treatment effect estimators as weighted outcomes, \url{https://arxiv.org/abs/2411.11559}.

Bach P, Kurz MS, Chernozhukov V, Spindler M, Klaassen S (2024). “DoubleML:
An Object-Oriented Implementation of Double Machine Learning in R.”
Journal of Statistical Software, 108(3), 1-56, \url{https://doi.org/10.18637/jss.v108.i03}.
}
\keyword{internal}
