% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml_with_smoother.R
\name{get_outcome_weights.dml_with_smoother}
\alias{get_outcome_weights.dml_with_smoother}
\title{Outcome weights for the \code{\link[grf]{causal_forest}} function}
\usage{
\method{get_outcome_weights}{dml_with_smoother}(object, ..., all_reps = FALSE)
}
\arguments{
\item{object}{An object of class \code{dml_with_smoother}, i.e. the result of running \code{\link{dml_with_smoother}}.}

\item{...}{Pass potentially generic \link{get_outcome_weights} options.}

\item{all_reps}{If \code{TRUE}, outcomes weights of each repetitions passed. Default \code{TRUE}.}
}
\value{
\itemize{
\item If \code{all_reps == TRUE}: Matrix of dimension (number of estimators x N) containing the outcome weights.
\item If \code{all_reps == FALSE}: list with two components:
\itemize{
\item omega: Matrix of dimension (number of estimators x N) containing the outcome weights.
\item omega_all_reps: A list containing the outcome weights of each repetition.
}
}
}
\description{
Post-estimation command to extract outcome weights for double ML
run with an outcome smoother.
}
