% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoubleML.R
\name{get_outcome_weights.DoubleML}
\alias{get_outcome_weights.DoubleML}
\title{Outcome weights for a \code{DoubleML} object}
\usage{
\method{get_outcome_weights}{DoubleML}(object, dml_data, ...)
}
\arguments{
\item{object}{A \code{DoubleML} object of class \code{DoubleMLPLR},
\code{DoubleMLPLIV}, \code{DoubleMLIIVM}, or \code{DoubleMLIRM}.}

\item{dml_data}{The \code{DoubleMLData} object used in estimation.}

\item{...}{Pass potentially generic \link{get_outcome_weights} options.}
}
\value{
An object of class \code{get_outcome_weights} containing:
\describe{
\item{omega}{Numeric vector of outcome weights of length \eqn{N}.}
\item{treat}{Numeric vector of treatment assignments.}
}
}
\description{
Post-estimation command to extract outcome weights for a
\code{DoubleML} object from the \pkg{DoubleML} package.
}
\details{
Outcome weights are available for \code{DoubleMLPLR}, \code{DoubleMLPLIV},
\code{DoubleMLIIVM}, and \code{DoubleMLIRM} models.
}
\examples{
\dontrun{
set.seed(123)

# Set the parameters
params <- list(alpha = 0, subsample = 1, max_delta_step = 0, base_score = 0)

# Define DML objects
ml_regr_xgb <- do.call(lrn, c(list("regr.xgboost"), params))
ml_l_plr <- ml_regr_xgb$clone()
ml_m_plr <- ml_regr_xgb$clone()
plr_data <- DoubleML::make_plr_CCDDHNR2018(500, dim_x = 10)
plr_obj <- DoubleML::DoubleMLPLR$new(plr_data, ml_l_plr, ml_m_plr, n_folds = 3)

# Fit the model
plr_obj$fit(store_models = TRUE, store_predictions = TRUE)

# Calculate the outcome weights
omega_plr <- get_outcome_weights(plr_obj, plr_data)

# Check equivalence 
all.equal(omega_plr$omega \%*\% plr_data$data$y, plr_obj$all_coef)
}

}
\references{
Knaus, M. C. (2024). Treatment effect estimators as weighted outcomes, \url{https://arxiv.org/abs/2411.11559}.

Bach P, Kurz MS, Chernozhukov V, Spindler M, Klaassen S (2024). “DoubleML:
An Object-Oriented Implementation of Double Machine Learning in R.”
Journal of Statistical Software, 108(3), 1-56, \url{https://doi.org/10.18637/jss.v108.i03}.
}
