% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/least_squares.R
\name{get_outcome_weights.ivreg}
\alias{get_outcome_weights.ivreg}
\title{Outcome weights for ivreg}
\usage{
\method{get_outcome_weights}{ivreg}(object, treat_pos = 2, ...)
}
\arguments{
\item{object}{An object of class \code{ivreg}, i.e. the result of running \code{\link[AER]{ivreg}}.}

\item{treat_pos}{The position of the treatment indicator in the model matrix. By default 2,
which is appropriate for the case with an intercept and the (endogneous) treatment indicator as the first explanatory variable.}

\item{...}{Pass potentially generic \link{get_outcome_weights} options.}
}
\value{
A vector of implied weights.
}
\description{
Post-estimation command to calculate the outcome weights for
Instrumental-Variable Regression implemented via the \code{\link[AER]{ivreg}} command.
Automatically takes sampling weights into account if used.
}
