% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/least_squares.R
\name{get_outcome_weights.lm}
\alias{get_outcome_weights.lm}
\title{Outcome weights for the lm command}
\usage{
\method{get_outcome_weights}{lm}(object, treat_pos = 2, ...)
}
\arguments{
\item{object}{An object of class \code{lm}, i.e. the result of running \code{\link[stats]{lm}}.}

\item{treat_pos}{The position of the treatment indicator in the model matrix. By default 2,
which is appropriate for the case with an intercept and the treatment indicator as the first explanatory variable.}

\item{...}{Pass potentially generic \link{get_outcome_weights} options.}
}
\value{
\link{get_outcome_weights} object with \code{omega} containing weights and \code{treat} the treatment
}
\description{
Post-estimation command to calculate the outcome weights for linear regression implemented via the \code{\link[stats]{lm}} command.
}
