% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outcome_weights.R
\name{get_outcome_weights}
\alias{get_outcome_weights}
\title{Outcome weights method}
\usage{
get_outcome_weights(object, ...)
}
\arguments{
\item{object}{An object, obtained from other packages.}

\item{...}{Additional arguments specific to object class implementations.
See the documentation which object requires which additional arguments.}
}
\value{
A list of at least these components:
\itemize{
\item omega: matrix (number of point estimates x number of estimation units) of outcome weights
\item treat: the treatment indicator to make it compatible with the cobalt package
}
}
\description{
This is a generic method for getting outcome weights.
It calculates the outcome weights for objects created by other packages.
See get_outcome_weight.<compatible_fct> in the package documentation for compatible functions.
}
\references{
Knaus, M. C. (2024). Treatment effect estimators as weighted outcomes, \url{https://arxiv.org/abs/2411.11559}.
}
