% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smoother_weights.R
\name{get_smoother_weights.cv.plasso}
\alias{get_smoother_weights.cv.plasso}
\title{Smoother weights for the \code{plasso::cv.plasso} function}
\usage{
\method{get_smoother_weights}{cv.plasso}(object, Xnew = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{cv.plasso}, i.e., the result of
calling \code{plasso::cv.plasso()}.}

\item{Xnew}{Optional covariate matrix of the test sample. If \code{NULL},
the smoother matrix for the training data is returned.}

\item{...}{Additional arguments passed to \link{get_smoother_weights}.}
}
\value{
An \eqn{N \times N} smoother matrix.
}
\description{
Post-estimation function to extract smoother weights for a post-Lasso regression
estimated via the \code{cv.plasso} function from the \pkg{plasso} package.
}
\examples{
\dontrun{
set.seed(123)
n <- 100
X <- matrix(rnorm(n * 3), ncol = 3)
Y <- 2 * X[, 1] + 3 * X[, 2] - 1 * X[, 3] + rnorm(n)

# Fit post-lasso model
object <- plasso::cv.plasso(x = X, y = Y)

# Get predictions
preds <- predict(object, newx = X, type = "response", s = "optimal", se_rule = 0)$plasso

# Get smoother matrix
S <- get_smoother_weights(object, Xnew = X)

# Check equivalence
all.equal(as.numeric(S \%*\% Y), as.numeric(preds))
}

}
\references{
Knaus, M. C. (2024). Treatment effect estimators as weighted outcomes, \url{https://arxiv.org/abs/2411.11559}.
}
