% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NuisanceParameters.R
\name{get_smoother_weights.ensemble_core}
\alias{get_smoother_weights.ensemble_core}
\title{Ensemble smoother weights}
\usage{
\method{get_smoother_weights}{ensemble_core}(object, methods, X_tr, X_te, quiet = FALSE, ...)
}
\arguments{
\item{object}{List of trained models from \code{\link[NuisanceParameters]{ensemble_core}}.}

\item{methods}{List of methods built via \code{\link[NuisanceParameters]{create_method}}.}

\item{X_tr}{Covariate matrix of training sample.}

\item{X_te}{Covariate matrix of test sample.}

\item{quiet}{If \code{FALSE}, the method currently computed is printed to the console.}

\item{...}{Ignore unused arguments.}
}
\value{
A three-dimensional array of dimension
\code{nrow(X_te)} × \code{nrow(X_tr)} × \code{length(methods)} containing smoother weights.
Each slice corresponds to one method, where each row gives the weights
assigned to training outcomes in predicting the respective test observation.
}
\description{
Extract smoother weights for a (new) covariate matrix based on the fully trained
models of an ensemble created by \code{\link[NuisanceParameters]{ensemble_core}}.
}
\keyword{internal}
