% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NuisanceParameters.R
\name{get_smoother_weights.knn_fit}
\alias{get_smoother_weights.knn_fit}
\title{Smoother weights from k-nearest neighbor algorithm (internal)}
\usage{
\method{get_smoother_weights}{knn_fit}(object, Xnew = NULL, ...)
}
\arguments{
\item{object}{Output of \code{\link[NuisanceParameters]{knn_fit}}.}

\item{Xnew}{Covariate matrix of test sample.
If not provided, prediction is done for the training sample.}

\item{...}{Ignore unused arguments.}
}
\value{
A matrix of smoother weights.
}
\description{
Extract smoother weights from a k-nearest neighbor algorithm.
Each weight equals \code{1 / k} for neighbors and 0 for non-neighbors.
}
\keyword{internal}
