% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smoother_weights.R
\name{get_smoother_weights.lm}
\alias{get_smoother_weights.lm}
\title{Smoother weights for the \code{stats::lm} function}
\usage{
\method{get_smoother_weights}{lm}(object, X, Xnew, ...)
}
\arguments{
\item{object}{An object of class \code{lm}, i.e., the result of calling \code{stats::lm()}.}

\item{X}{Covariate matrix of the training sample.}

\item{Xnew}{Optional covariate matrix of the test sample. If \code{NULL},
the smoother matrix for the training data is returned.}

\item{...}{Additional arguments passed to \link{get_smoother_weights}.}
}
\value{
An \eqn{N \times N} smoother matrix.
}
\description{
Post-estimation function to extract smoother weights for a linear regression
estimated via the \code{lm} function from the \pkg{stats} package.
}
\examples{
\dontrun{
set.seed(123)
n <- 100
X <- matrix(rnorm(n * 3), ncol = 3)
Y <- 2 * X[, 1] + 3 * X[, 2] - 1 * X[, 3] + rnorm(n)

# Fit linear model
object <- lm(Y ~ 0 + X)

# Get fitted values
preds <- object$fitted.values

# Get smoother matrix
S <- get_smoother_weights(object, X = X, Xnew = X)

# Check equivalence
all.equal(as.numeric(S \%*\% Y), as.numeric(preds))
}

}
