% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NuisanceParameters.R
\name{get_smoother_weights.mean_fit}
\alias{get_smoother_weights.mean_fit}
\title{Compute smoother weights from arithmetic mean}
\usage{
\method{get_smoother_weights}{mean_fit}(object, Xnew = NULL, ...)
}
\arguments{
\item{object}{An object returned by \code{\link[NuisanceParameters]{mean_fit}}.}

\item{Xnew}{Optional covariate matrix of the test sample.
If not provided, predictions are computed for the training sample.}

\item{...}{Ignored arguments.}
}
\value{
A matrix of smoother weights.
}
\description{
Generate smoother weights for a test sample using arithmetic mean fitting
of the training sample.
}
\keyword{internal}
