% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smoother_weights.R
\name{get_smoother_weights.ranger}
\alias{get_smoother_weights.ranger}
\title{Smoother weights for the \code{ranger::ranger} function}
\usage{
\method{get_smoother_weights}{ranger}(object, X = NULL, Xnew = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{ranger}, i.e., the result of
calling \code{ranger::ranger()}.}

\item{X}{Optional covariate matrix of the training sample;
used if not stored in the fitted object.}

\item{Xnew}{Covariate matrix of test sample.
If not provided, prediction is done for the training sample.}

\item{...}{Additional arguments passed to \link{get_smoother_weights}.}
}
\value{
An \eqn{N \times N} smoother matrix.
}
\description{
Post-estimation function to extract smoother weights for a random forest model
estimated via the \code{ranger} function from the \pkg{ranger} package.
}
\examples{
\dontrun{
set.seed(123)
n <- 100
X <- matrix(rnorm(n * 3), ncol = 3, dimnames = list(NULL, 1:3))
Y <- 2 * X[, 1] + 3 * X[, 2] - 1 * X[, 3] + rnorm(n)

# Fit random forest
object <- ranger::ranger(x = X, y = Y, keep.inbag = TRUE)

# Get predictions
preds <- predict(object, data = X)$predictions

# Get smoother matrix
S <- get_smoother_weights(object, X = X)

# Check equivalence
all.equal(as.numeric(S \%*\% Y), as.numeric(preds))
}

}
