% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smoother_weights.R
\name{get_smoother_weights}
\alias{get_smoother_weights}
\title{Smoother weights method}
\usage{
get_smoother_weights(object, ...)
}
\arguments{
\item{object}{An object obtained from other packages.}

\item{...}{Additional arguments specific to object class implementations.
See the documentation for which objects require which additional arguments.}
}
\value{
An \eqn{N \times N} smoother matrix.
}
\description{
This is a generic method for getting smoother weights.
It calculates the smoother weights for objects created by other packages.
See get_smoother_weight.<compatible_class> in the package documentation for compatible functions.
}
