% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smoother_weights.R
\name{get_smoother_weights.regression_forest}
\alias{get_smoother_weights.regression_forest}
\title{Smoother weights for the \code{grf::regression_forest} function}
\usage{
\method{get_smoother_weights}{regression_forest}(object, Xnew = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{regression_forest}, i.e.,
the result of calling \code{grf::regression_forest()}.}

\item{Xnew}{Optional covariate matrix of the test sample. If \code{NULL},
the smoother matrix for the training data is returned.}

\item{...}{Additional arguments passed to \link{get_smoother_weights}.}
}
\value{
An \eqn{N \times N} smoother matrix.
}
\description{
Post-estimation function to extract smoother weights for a random forest model
estimated via the \code{regression_forest} function from the \pkg{grf} package.
}
\examples{
\dontrun{
set.seed(123)
n <- 100
X <- matrix(rnorm(n * 3), ncol = 3)
Y <- 2 * X[, 1] + 3 * X[, 2] - 1 * X[, 3] + rnorm(n)

# Fit regression forest
object <- grf::regression_forest(X = X, Y = Y)

# Get predictions
preds <- predict(object, newdata = X)$predictions

# Get smoother matrix
S <- get_smoother_weights(object)

# Check equivalence
all.equal(as.numeric(S \%*\% Y), as.numeric(preds))
}

}
\references{
Athey, S., Tibshirani, J., & Wager, S. (2019). Generalized random forest. The Annals of Statistics, 47(2), 1148-1178.

Knaus, M. C. (2024). Treatment effect estimators as weighted outcomes, \url{https://arxiv.org/abs/2411.11559}.
}
