% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NuisanceParameters.R
\name{get_smoother_weights.ridge_fit}
\alias{get_smoother_weights.ridge_fit}
\title{Smoother weights for ridge regression (internal)}
\usage{
\method{get_smoother_weights}{ridge_fit}(object, Xnew = NULL, ...)
}
\arguments{
\item{object}{An object returned by \code{\link[NuisanceParameters]{ridge_fit}}.}

\item{Xnew}{Optional covariate matrix of the test sample. If \code{NULL},
the smoother matrix for the training data is returned. If standardization
is used, must be standardized using the same center and scale as \code{X}.}

\item{...}{Additional arguments passed to \link{get_smoother_weights}.}
}
\value{
An \eqn{N \times N} smoother matrix.
}
\description{
Post-estimation function to extract smoother weights for a ridge regression.
The models are estimated via a wrapper of \code{cv.glmnet} in the
\pkg{NuisanceParameters} package with \code{alpha = 0}.
}
\keyword{internal}
