% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smoother_weights.R
\name{get_smoother_weights.rlasso}
\alias{get_smoother_weights.rlasso}
\title{Smoother weights for the \code{hdm::rlasso} function}
\usage{
\method{get_smoother_weights}{rlasso}(object, Xnew = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{rlasso}, i.e., the result of
calling \code{hdm::rlasso()}.}

\item{Xnew}{Optional covariate matrix of the test sample. If \code{NULL},
the smoother matrix for the training data is returned.}

\item{...}{Additional arguments passed to \link{get_smoother_weights}.}
}
\value{
An \eqn{N \times N} smoother matrix.
}
\description{
Post-estimation function to extract smoother weights for a post-Lasso regression
estimated via the \code{rlasso} function from the \pkg{hdm} package.
}
\examples{
\dontrun{
set.seed(123)
n <- 100
X <- matrix(rnorm(n * 3), ncol = 3)
Y <- 2 * X[, 1] + 3 * X[, 2] - 1 * X[, 3] + rnorm(n)

# Fit post-lasso model
object <- hdm::rlasso(x = X, y = Y, post = TRUE)

# Get predictions
preds <- predict(object, newdata = X)

# Get smoother matrix
S <- get_smoother_weights(object, Xnew = X)

# Check equivalence
all.equal(as.numeric(S \%*\% Y), as.numeric(preds))
}

}
\references{
Knaus, M. C. (2024). Treatment effect estimators as weighted outcomes, \url{https://arxiv.org/abs/2411.11559}.
}
