% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NuisanceParameters.R
\name{get_smoother_weights.xgboost_fit}
\alias{get_smoother_weights.xgboost_fit}
\title{Smoother weights for the \code{xgboost_fit} function}
\usage{
\method{get_smoother_weights}{xgboost_fit}(object, Xnew = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{xgboost_fit}, i.e., the result of calling \code{xgboost_fit}.}

\item{Xnew}{Optional covariate matrix of the test sample. If \code{NULL},
the smoother matrix for the training data is returned.}

\item{...}{Additional arguments passed to \link{get_smoother_weights}.}
}
\value{
An \eqn{N \times N} smoother matrix.
}
\description{
Post-estimation function to extract smoother weights for an XGBoost model
estimated via the \code{xgboost_fit} function from the \pkg{NuisanceParameters} package.
}
\keyword{internal}
