\name{OutcomeWeights-package}
\alias{OutcomeWeights-package}
\alias{OutcomeWeights}
\docType{package}
\title{
\packageTitle{OutcomeWeights}
}
\description{
\packageDescription{OutcomeWeights}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{OutcomeWeights}

Maintainer: \packageMaintainer{OutcomeWeights}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = OutcomeWeights)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{OutcomeWeights}}
%\section{Documentation Index}{\packageIndices{OutcomeWeights}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
