% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outcome_weights.R
\name{pive_weight_maker}
\alias{pive_weight_maker}
\title{Outcome weights maker for pseudo-IV estimators.}
\usage{
pive_weight_maker(Z.tilde, D.tilde, T_mat)
}
\arguments{
\item{Z.tilde}{Numeric vector of pseudo-instrument outcomes.}

\item{D.tilde}{Numeric vector of pseudo-treatment.}

\item{T_mat}{Transformation matrix}
}
\value{
A vector of outcome weights.
}
\description{
This is a generic function taking pseudo-instrument,
pseudo-treatment and the transformation matrix as inputs and returning
outcome weights
}
\references{
Knaus, M. C. (2024). Treatment effect estimators as weighted outcomes, soon on 'arXiv'.
}
