% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml_with_smoother.R
\name{plot.dml_with_smoother}
\alias{plot.dml_with_smoother}
\title{\code{plot} method for class \code{\link{dml_with_smoother}}}
\usage{
\method{plot}{dml_with_smoother}(x, ..., alpha = 0.05, contrast = FALSE)
}
\arguments{
\item{x}{Object of class \code{\link{dml_with_smoother}}.}

\item{...}{Pass generic \code{\link[base]{plot}} options.}

\item{alpha}{Significance level for confidence intervals (default 0.05).}

\item{contrast}{Shows the differences between the coefficients.}
}
\value{
ggplot with point estimates and confidence intervals.
}
\description{
\code{plot} method for class \code{\link{dml_with_smoother}}
}
