% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NuPa_generic.R
\name{prep_cf_mat}
\alias{prep_cf_mat}
\title{Creates matrix of binary cross-fitting fold indicators (N x # cross-folds)}
\usage{
prep_cf_mat(n, cf, w_mat = NULL, cl = NULL)
}
\arguments{
\item{n}{Number of observations.}

\item{cf}{Number of cross-fitting folds.}

\item{w_mat}{Optional logical matrix of treatment indicators (N x T+1).
If specified, cross-fitting folds will preserve the treatment ratios from full sample.}

\item{cl}{Optional vector of cluster variable if cross-fitting should account for clusters.}
}
\value{
Logical matrix of cross-fitting folds (N x # folds).
}
\description{
Creates matrix of binary cross-fitting fold indicators (N x # cross-folds)
}
