% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{summarize_weights}
\alias{summarize_weights}
\title{Calls C++ implementation to produce summary statistics of the outcome weights.}
\usage{
summarize_weights(omega, treat)
}
\arguments{
\item{omega}{Outcome weights matrix with dimension number of weight vectors for which balancing should be checked
x number of training units.}

\item{treat}{Binary treatment variable.}
}
\value{
3D-array of dimension
\itemize{
\item c("Control","Treated") x
\item number of training units x
\item c("Minimum weight","Maximum weight","\% Negative","Sum largest 10\%","Sum of weights","Sum of absolute weights")#'
}
}
\description{
Calculates several summary measures of potentially many outcome weights.
}
