% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml_with_smoother.R
\name{summary.dml_with_smoother}
\alias{summary.dml_with_smoother}
\title{\code{summary} method for class \code{\link{dml_with_smoother}}}
\usage{
\method{summary}{dml_with_smoother}(object, contrast = FALSE, quiet = FALSE, ...)
}
\arguments{
\item{object}{Object of class \code{\link{dml_with_smoother}}.}

\item{contrast}{Tests the differences between the coefficients.}

\item{quiet}{If TRUE, results are passed but not printed.}

\item{...}{further arguments passed to \code{printCoefmat}}
}
\value{
Invisible matrix with estimator(s) in the rows and
c("Estimate","SE","t","p") in the columns.
}
\description{
\code{summary} method for class \code{\link{dml_with_smoother}}
}
