% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outcome_weights.R
\name{summary.get_outcome_weights}
\alias{summary.get_outcome_weights}
\title{\code{summary} method for class \code{outcome_weights}}
\usage{
\method{summary}{get_outcome_weights}(object, quiet = FALSE, digits = 4, epsilon = 1e-04, ...)
}
\arguments{
\item{object}{\link{get_outcome_weights} object.}

\item{quiet}{If TRUE, results are passed but not printed.}

\item{digits}{Number of digits to be displayed. Default 4.}

\item{epsilon}{Threshold below which in absolute values non-zero but small values should be displayed as < ...}

\item{...}{further arguments passed to \code{printCoefmat}}
}
\value{
3D-array of dimension
\itemize{
\item c("Control","Treated") x
\item number of point estimates x
\item c("Minimum weight","Maximum weight","\% Negative","Sum largest 10\%","Sum of weights","Sum of absolute weights")
}
}
\description{
Calculates several summary measures of potentially many outcome weights.
}
