% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{summary.standardized_mean_differences}
\alias{summary.standardized_mean_differences}
\title{\code{summary} method for class \code{\link{standardized_mean_differences}}}
\usage{
\method{summary}{standardized_mean_differences}(object, ...)
}
\arguments{
\item{object}{Object of class \code{\link{standardized_mean_differences}}.}

\item{...}{further arguments passed to summary method.}
}
\value{
3D-array of dimension
\itemize{
\item c("Maximum absolute SMD","Mean absolute SMD","Median absolute SMD", / \% of absolute SMD > 20",
"# / \% of absolute SMD > 10","# / \% of absolute SMD > 5") x
\item c("Balancing","Targeting") x
\item number of weight vectors for which balancing should be checked
}
}
\description{
Calls a C++ function to quickly summarize potentially many standardized mean differences.
}
\references{
Rosenbaum, P. R., & Rubin, D. B. (1984). Reducing bias in observational studies using subclassification on the propensity score.
Journal of the American Statistical Association, 79 (387), 516–524.
}
