% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml_with_smoother.R
\name{get_outcome_weights.dml_with_smoother}
\alias{get_outcome_weights.dml_with_smoother}
\title{Outcome weights for the \code{\link{dml_with_smoother}} function}
\usage{
\method{get_outcome_weights}{dml_with_smoother}(object, ..., all_reps = FALSE)
}
\arguments{
\item{object}{An object of class \code{dml_with_smoother}, i.e. the result of running \code{\link{dml_with_smoother}}.}

\item{...}{Pass potentially generic \link{get_outcome_weights} options.}

\item{all_reps}{If \code{TRUE}, outcomes weights of each repetitions passed. Default \code{FALSE}.}
}
\value{
\itemize{
\item If \code{all_reps == FALSE}: \link{get_outcome_weights} object
\item If \code{all_reps == TRUE}: additionally list \code{omega_all_reps}:
A list containing the outcome weights of each repetition.
}
}
\description{
Post-estimation command to extract outcome weights for double ML
run with an outcome smoother.
}
\references{
Knaus, M. C. (2024). Treatment effect estimators as weighted outcomes, \url{https://arxiv.org/abs/2411.11559}.
}
