% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutcomeWeights-package.R
\docType{package}
\name{OutcomeWeights-package}
\alias{OutcomeWeights}
\alias{OutcomeWeights-package}
\title{OutcomeWeights: Outcome Weights of Treatment Effect Estimators}
\description{
Many treatment effect estimators can be written as weighted outcomes. These weights have established use cases like checking covariate balancing via packages like 'cobalt'. This package takes the original estimator objects and outputs these outcome weights. It builds on the general framework of Knaus (2024) \doi{10.48550/arXiv.2411.11559}. This version is compatible with the 'grf' package and provides an internal implementation of Double Machine Learning.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MCKnaus/OutcomeWeights}
  \item Report bugs at \url{https://github.com/MCKnaus/OutcomeWeights/issues}
}

}
\author{
\strong{Maintainer}: Michael C. Knaus \email{michael.knaus@uni-tuebingen.de} (\href{https://orcid.org/0000-0002-7328-1363}{ORCID})

Other contributors:
\itemize{
  \item Henri Pfleiderer \email{henri.1999@web.de} [contributor]
}

}
\keyword{internal}
